requirements.txt
boto3~=1.0
botocore~=1.0
jmespath~=0.10.0
python-dateutil~=2.0
s3transfer~=0.5.0
six~=1.0
urllib3~=1.0
PyYaml~=6.0